requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud and Firebase
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.18.1
google-cloud-scheduler==2.16.1
google-cloud-logging==3.8.0
firebase-admin==6.4.0

# AI and ML
google-generativeai==0.3.2

# Social Media APIs
praw==7.7.1
tweepy==4.14.0
linkedin-api==2.2.0

# HTTP clients and utilities
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data processing
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3

# Environment and configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-test==0.1.7

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Job scheduling
celery==5.3.4
redis==5.0.1